# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Charity Subscription Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  /** @nullable */
  charityId?: number | null;
  /** @nullable */
  charityPercent?: number | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UpdateUserBody {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Subscription {
  id: number;
  userId: number;
  plan: string;
  status: string;
  amount: number;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  renewalDate?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  createdAt: string;
}

export interface UserWithSubscription {
  id: number;
  email: string;
  name: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
  subscription?: Subscription | null;
}

export interface CreateSubscriptionBody {
  plan: string;
  /** @nullable */
  paymentMethodId?: string | null;
}

export interface Score {
  id: number;
  userId: number;
  value: number;
  scoreDate: string;
  createdAt: string;
}

export interface AddScoreBody {
  value: number;
  scoreDate: string;
}

export interface UpdateScoreBody {
  value?: number;
  scoreDate?: string;
}

export interface Charity {
  id: number;
  name: string;
  description: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  featured: boolean;
  totalContributions: number;
  createdAt: string;
}

export interface CreateCharityBody {
  name: string;
  description: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  featured?: boolean;
}

export interface UpdateCharityBody {
  name?: string;
  description?: string;
  category?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  featured?: boolean;
}

export interface UserCharity {
  id: number;
  userId: number;
  charityId: number;
  charityPercent: number;
  charity: Charity;
  createdAt: string;
}

export interface SetCharitySelectionBody {
  charityId: number;
  charityPercent: number;
}

export interface Draw {
  id: number;
  month: number;
  year: number;
  status: string;
  prizePool: number;
  jackpotRollover: number;
  /** @nullable */
  drawnNumbers?: string | null;
  algorithm: string;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
}

export interface CreateDrawBody {
  month: number;
  year: number;
  algorithm?: string;
  jackpotRollover?: number;
}

export interface DrawResult {
  id: number;
  drawId: number;
  userId: number;
  matchTier: number;
  prizeAmount: number;
  userName: string;
  userEmail: string;
}

export interface DrawWithResults {
  id: number;
  month: number;
  year: number;
  status: string;
  prizePool: number;
  jackpotRollover: number;
  /** @nullable */
  drawnNumbers?: string | null;
  algorithm: string;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  results: DrawResult[];
}

export interface DrawSimulationResult {
  drawnNumbers: number[];
  tier5Winners: DrawResult[];
  tier4Winners: DrawResult[];
  tier3Winners: DrawResult[];
  prizePool: number;
}

export interface Winner {
  id: number;
  drawId: number;
  userId: number;
  matchTier: number;
  prizeAmount: number;
  status: string;
  /** @nullable */
  proofUrl?: string | null;
  /** @nullable */
  adminNote?: string | null;
  userName: string;
  userEmail: string;
  drawMonth: number;
  drawYear: number;
  createdAt: string;
}

export interface UploadProofBody {
  proofUrl: string;
}

export interface VerifyWinnerBody {
  approved: boolean;
  /** @nullable */
  adminNote?: string | null;
}

export interface Analytics {
  totalUsers: number;
  activeSubscribers: number;
  totalPrizePool: number;
  totalCharityContributions: number;
  totalDraws: number;
  totalWinners: number;
  monthlyRevenue: number;
  pendingPayouts: number;
}

export interface PrizePool {
  currentPool: number;
  jackpotRollover: number;
  tier5Prize: number;
  tier4Prize: number;
  tier3Prize: number;
  activeSubscribers: number;
  monthlyContribution: number;
}

export type ListUsersParams = {
  search?: string;
  status?: string;
};

export type ListCharitiesParams = {
  search?: string;
  category?: string;
};