# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Golf Charity Subscription Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Users
...
  /charities:
    get:
      operationId: listCharities
      tags: [charities]
      summary: List all charities
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of charities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Charity"

    post:
      operationId: createCharity
      tags: [charities]
      summary: Create a charity (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharityBody"
      responses:
        "201":
          description: Charity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charity"
...
  /charities/{id}:
    get:
      operationId: getCharity
      tags: [charities]
      summary: Get charity by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Charity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charity"
        "404":
          description: Not found

    patch:
      operationId: updateCharity
      tags: [charities]
      summary: Update a charity (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCharityBody"
      responses:
        "200":
          description: Charity updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Charity"

    delete:
      operationId: deleteCharity
      tags: [charities]
      summary: Delete a charity (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /charities/my-selection:
    get:
      operationId: getMyCharitySelection
      tags: [charities]
      summary: Get current user's charity selection
      responses:
        "200":
          description: User charity selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCharity"
        "404":
          description: No selection

    post:
      operationId: setMyCharitySelection
      tags: [charities]
      summary: Set current user's charity selection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCharitySelectionBody"
      responses:
        "200":
          description: Selection updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCharity"

  # Draws
...
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        charityId:
          type: ["integer", "null"]
        charityPercent:
          type: ["integer", "null"]
      required:
        - email
        - password
        - name
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
...
    Charity:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        totalContributions:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - category
        - featured
        - totalContributions
        - createdAt
    CreateCharityBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        featured:
          type: boolean
      required:
        - name
        - description
        - category
    UpdateCharityBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        featured:
          type: boolean
    UserCharity:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        charityId:
          type: integer
        charityPercent:
          type: integer
        charity:
          $ref: "#/components/schemas/Charity"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - charityId
        - charityPercent
        - charity
        - createdAt
    SetCharitySelectionBody:
      type: object
      properties:
        charityId:
          type: integer
        charityPercent:
          type: integer
      required:
        - charityId
        - charityPercent
...