# lib/api-zod/src/generated/types/createCharityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Charity Subscription Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCharityBody {
  name: string;
  description: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  featured?: boolean;
}